Add Gobro user creation to mosquitto script
Updated script to create a user for Gobro with provided credentials.

+++ create-mosquitto-users.sh
@@ -5,11 +5,16 @@ PASSWORD_FILE="/mosquitto/config/passwordfile"
 
 # Check if 
 : "${GROWATT_SN:?Serial not set}"
+: "${GROBRO_USER:?GroBro User not set}"
+: "${GROBRO_PASS:?GroBro Pass not set}"
 
+# Clear Password File
 rm -f "$PASSWORD_FILE"
 touch "$PASSWORD_FILE"
 
-# Reas Serials to users
+# Create Gobro User
+mosquitto_passwd -b "$PASSWORD_FILE" "$GROBRO_USER" "$GROBRO_PASS"
+# Read Serials to and Create Users
 IFS=',' read -ra SERIALS <<< "$GROWATT_SN"
 
 for serial in "${SERIALS[@]}"; do
